Move login to user
